seperate mac and windows voice
